+++ src/main/java/macaca/java/biz/BaseMacacaClient.java
@@ -6,6 +6,7 @@ import java.math.BigInteger;
 import java.security.MessageDigest;
 import java.sql.Driver;
 
+import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 
 import macaca.client.MacacaClient;
@@ -454,6 +455,125 @@ public class BaseMacacaClient extends MacacaClient {
 
    	}
 
+   	/**
+   	 * 滑动到最底部
+   	 */
+   	public void scrollToBottom () {
+   		JSONObject windowSize;
+   		try {
+   			windowSize = getWindowSize();
+   			int windowWidth = windowSize.getIntValue("width");
+   			int windowHeight = windowSize.getIntValue("height");
+
+   			int startX = windowWidth-20;
+   			int endX = startX;
+   			int startY = windowHeight*3/5;
+   			int endY = windowHeight*2/5;
+
+   			String beforeScreenShot = null ;
+   			String afterScreenShot = null;
+   			String beforePng = "before_bottom.png";
+   			String afterPng = "after_bottom.png";
+   			int flag = 10;
+   			while (flag > 0) {
+
+   				File shotOne = new File(beforePng);
+   				File shotTwo = new File(afterPng);
+   				beforeScreenShot = BaseUtils.getFileMD5(shotOne);
+   				afterScreenShot = BaseUtils.getFileMD5(shotTwo);
+   				if (beforeScreenShot != null &&
+   					beforeScreenShot.length() > 0) {
+   					if (beforeScreenShot.equals(afterScreenShot)) {
+   						// the same screen image ,it means current view has scroll to bottom
+   						System.out.println("scroollToBottom");
+   						deleteDiffImages();
+   						return;
+   					}
+   				}
+
+   				saveScreenshot(beforePng);
+   				System.out.println("scroll: ("+startX+","+startY+","+endX+","+endY+")");
+   				drag(startX, startY, endX, endY, 0.05,10);
+   				Thread.sleep(1000);
+   				saveScreenshot(afterPng);
+   				// 为防止死循环，最多滑动10次
+   				flag--;
+
+   			}
+
+
+   			deleteDiffImages();
+
+
+   		} catch (Exception e) {
+   			// TODO Auto-generated catch block
+   			deleteDiffImages();
+   			e.printStackTrace();
+   		}
+
+   		deleteDiffImages();
+   	}
+
+	/**
+   	 * 滑动到最顶部
+   	 */
+   	public void scrollToTop () {
+   		JSONObject windowSize;
+   		try {
+   			windowSize = getWindowSize();
+   			int windowWidth = windowSize.getIntValue("width");
+   			int windowHeight = windowSize.getIntValue("height");
+
+   			int startX = windowWidth-20;
+   			int endX = startX;
+   			int startY = windowHeight*2/5;
+   			int endY = windowHeight*3/5;
+
+   			String beforeScreenShot = null ;
+   			String afterScreenShot = null;
+   			String beforePng = "before_top.png";
+   			String afterPng = "after_top.png";
+   			int flag = 10;
+   			while (flag > 0) {
+
+   				File shotOne = new File(beforePng);
+   				File shotTwo = new File(afterPng);
+   				beforeScreenShot = BaseUtils.getFileMD5(shotOne);
+   				afterScreenShot = BaseUtils.getFileMD5(shotTwo);
+   				if (beforeScreenShot != null &&
+   					beforeScreenShot.length() > 0) {
+   					if (beforeScreenShot.equals(afterScreenShot)) {
+   						// the same screen image ,it means current view has scroll to bottom
+   						System.out.println("scroollToBottom");
+   						deleteDiffImages();
+   						return;
+   					}
+   				}
+
+   				saveScreenshot(beforePng);
+   				System.out.println("scroll: ("+startX+","+startY+","+endX+","+endY+")");
+   				drag(startX, startY, endX, endY, 0.05,10);
+   				Thread.sleep(1000);
+   				saveScreenshot(afterPng);
+   				// 为防止死循环，最多滑动10次
+   				flag--;
+
+   			}
+
+
+   			deleteDiffImages();
+
+
+   		} catch (Exception e) {
+   			// TODO Auto-generated catch block
+   			deleteDiffImages();
+   			e.printStackTrace();
+   		}
+
+   		deleteDiffImages();
+   	}
+
+
    	private void deleteDiffImages() {
    		try {
    			// 如果存在图片，则删除,防止污染图片
@@ -515,6 +635,29 @@ public class BaseMacacaClient extends MacacaClient {
  	}
 
 
+ 	 // switch to the context of the last pushed webview
+ 	/**
+ 	 * 从native切换到webview
+ 	 * @param driver
+ 	 * @return
+ 	 * @throws Exception
+ 	 */
+    public  void switchFromNativeToWebView() throws Exception {
+        JSONArray contexts = contexts();
+        this.context(contexts.get(contexts.size() - 1).toString());
+    }
+
+
+    /**
+     * 从webview切换到native
+     * @param driver
+     * @return
+     * @throws Exception
+     */
+    public  void switchFromeWebviewToNative() throws Exception {
+        JSONArray contexts = contexts();
+        context(contexts.get(0).toString());
+    }